slowconv2d fixed + fwd/bwd tests added

+++ nn/variable.py
@@ -2,6 +2,6 @@ import numpy as np
 
 
 class Variable:
-    def __init__(self, data):
-        self.data = data
-        self.grad = np.zeros_like(data)
+    def __init__(self, data, dtype=np.float32):
+        self.data = data.astype(dtype)
+        self.grad = np.zeros_like(data, dtype=dtype)
